+++ src/main/java/agents/flexibility/Strategist.java
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2024 German Aerospace Center <amiris@dlr.de>
+// SPDX-FileCopyrightText: 2025 German Aerospace Center <amiris@dlr.de>
 //
 // SPDX-License-Identifier: Apache-2.0
 package agents.flexibility;
@@ -22,16 +22,25 @@ import de.dlr.gitlab.fame.time.TimeStamp;
 
 /** Base class for strategists that operate some kind of flexibility, e.g., an energy storage or flexible electrolysis
  * 
- * @author Christoph Schimeczek */
+ * @author Christoph Schimeczek, Felix Nitsch */
 public abstract class Strategist {
 	/** Error message if used {@link Strategist} type cannot provide forecasts */
 	protected static final String ERR_PROVIDE_FORECAST = "Cannot provide bid forecasts with flexibility strategist of type: ";
-	/** Error message  if used {@link Strategist} type cannot deal with incoming forecasts */
+	/** Error message if used {@link Strategist} type cannot deal with incoming forecasts */
 	protected static final String ERR_USE_PRICE_FORECAST = "Cannot use price forecasts with flexibility strategist of type: ";
-	/** Error message  if used {@link Strategist} type cannot process merit order forecast */
+	/** Error message if used {@link Strategist} type cannot process merit order forecast */
 	protected static final String ERR_USE_MERIT_ORDER_FORECAST = "Cannot use merit order forecasts with flexibility strategist of type: ";
-	/** Error message  if used {@link Strategist} type is not implemented */
+	/** Error message if used {@link Strategist} type is not implemented */
 	protected static final String ERR_UNKNOWN_STRATEGIST = "This type of flexibility strategist is not implemented: ";
+	/** Error message if used {@link ForecastUpdateType} is not implemented */
+	protected static final String ERR_UNKNOWN_UPDATE_TYPE = "This type of forecast update is not implemented: ";
+
+	private static enum ForecastUpdateType {
+		/** Forecasts are requested for all time steps, discarding previously received electricity price forecasts. */
+		ALL,
+		/** Forecasts are only requested for missing time steps, therefore updated incrementally. */
+		INCREMENTAL,
+	};
 
 	/** Hard coded time granularity of {@link Strategist} */
 	public final static TimeSpan OPERATION_PERIOD = new TimeSpan(1, Interval.HOURS);
@@ -42,6 +51,8 @@ public abstract class Strategist {
 	protected final int scheduleDurationPeriods;
 	/** safety margins at bidding */
 	private final double bidTolerance;
+	/** forecast update type */
+	private final ForecastUpdateType forecastUpdateType;
 
 	/** schedule for the electricity demand (or charging) schedule */
 	protected double[] demandScheduleInMWH;
@@ -58,6 +69,10 @@ public abstract class Strategist {
 	public static final ParameterBuilder scheduleDurationParam = Make.newInt("ScheduleDurationInHours");
 	/** Strategist input parameter: safety margin at bidding */
 	public static final ParameterBuilder bidToleranceParam = Make.newDouble("BidToleranceInEURperMWH").optional();
+	/** Strategist input parameter: forecast update type */
+	public static final ParameterBuilder forecastUpdateTypeParam = Make
+			.newEnum("ForecastUpdateType", ForecastUpdateType.class)
+			.optional();
 
 	/** Creates new Strategist based on the given input
 	 * 
@@ -67,6 +82,8 @@ public abstract class Strategist {
 		forecastSteps = input.getInteger("ForecastPeriodInHours");
 		scheduleDurationPeriods = input.getInteger("ScheduleDurationInHours");
 		bidTolerance = input.getDoubleOrDefault("BidToleranceInEURperMWH", 1E-3);
+		forecastUpdateType = input.getEnumOrDefault("ForecastUpdateType", ForecastUpdateType.class,
+				ForecastUpdateType.INCREMENTAL);
 		allocateSchedulingArrays();
 	}
 
@@ -100,20 +117,29 @@ public abstract class Strategist {
 		return getMissingForecastTimes(sensitivities, firstTime);
 	}
 
-	/** Returns list of times at which given TreeMap misses entries needed for schedule planning
+	/** Returns list of times at which given TreeMap are requested for schedule planning
 	 * 
 	 * @param map to be inspected for existing TimeSegment keys
 	 * @param firstTime first time period to be covered by a created schedule
-	 * @return List of {@link TimeStamp}s at which given map misses entries */
+	 * @return List of {@link TimeStamp}s at which entries are requested */
 	protected ArrayList<TimeStamp> getMissingForecastTimes(TreeMap<TimePeriod, ?> map, TimePeriod firstTime) {
-		ArrayList<TimeStamp> missingTimes = new ArrayList<>();
+		ArrayList<TimeStamp> requestedTimes = new ArrayList<>();
 		for (int period = 0; period < forecastSteps; period++) {
 			TimePeriod timeSegment = firstTime.shiftByDuration(period);
-			if (!map.containsKey(timeSegment)) {
-				missingTimes.add(timeSegment.getStartTime());
+			switch (forecastUpdateType) {
+				case INCREMENTAL:
+					if (!map.containsKey(timeSegment)) {
+						requestedTimes.add(timeSegment.getStartTime());
+					}
+					break;
+				case ALL:
+					requestedTimes.add(timeSegment.getStartTime());
+					break;
+				default:
+					throw new RuntimeException(ERR_UNKNOWN_UPDATE_TYPE + forecastUpdateType);
 			}
 		}
-		return missingTimes;
+		return requestedTimes;
 	}
 
 	/** Stores given supply and demand bid forecasts for the associated TimePeriod: merit-order forecasting method
@@ -134,7 +160,7 @@ public abstract class Strategist {
 
 	/** optional action called on given MeritOrderSensitivities
 	 * 
-	 * @param sensitivity to be modified 
+	 * @param sensitivity to be modified
 	 * @param timePeriod that the sensitivity is valid for */
 	protected void callOnSensitivity(MeritOrderSensitivity sensitivity, TimePeriod timePeriod) {};